requirements.txt
tweepy
python-dotenv
mysql-connector-python
mecab-python3
pytz
